LibWeb: Split ResolutionStyleValue out of StyleValue.{h,cpp}
